move setAxisOrientations into optionsService
