+++ projects/ngx-d3timeline/src/lib/options.service.ts
@@ -1,9 +1,15 @@
 import { Injectable } from '@angular/core';
 import { Orientation } from './orientation';
 import { TimelineView } from './view/timeline-view';
+import { AxisOrientations } from './axis-orientations';
 
 @Injectable({ providedIn: 'root' })
 export class OptionsService {
+  setAxisOrientations(timeOrientation: Orientation): AxisOrientations {
+    const resourceOrientation = this.flipOrientation(timeOrientation);
+    return { time: timeOrientation, resource: resourceOrientation };
+  }
+
   getTranslation(
     range: number,
     orientation: Orientation,
@@ -12,5 +18,11 @@ export class OptionsService {
     return orientation === Orientation.Vertical
       ? `translate(${timelineView.left}, ${range})`
       : `translate(${range}, ${timelineView.top})`;
+  }
+
+  private flipOrientation(orientation: Orientation) {
+    return orientation === Orientation.Vertical
+      ? Orientation.Horizontal
+      : Orientation.Vertical;
   }
 }
